pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "digital-necrosis"
version = "0.1.0"
description = "Irreversible Adaptive Memory Precision Under Budget Constraints"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Ramchand", email = "ramchand@murailabs.org"},
]
keywords = [
    "LLM agents",
    "long-term memory",
    "selective forgetting",
    "adaptive compression",
    "vector quantization",
    "alignment drift",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "chromadb>=0.5.0,<0.6.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "sentence-transformers>=3.0",
    "torch>=2.3",
    "transformers>=4.40",
]

[project.optional-dependencies]
gpu = [
    "vllm>=0.6.0",
    "auto-gptq>=0.7",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
analysis = [
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "statsmodels>=0.14",
    "scikit-learn>=1.5",
]

[project.urls]
Repository = "https://github.com/murai-labs/digital-necrosis"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
